+++ lib/src/posthog_flutter_platform_interface.dart
@@ -44,6 +44,7 @@ abstract class PosthogFlutterPlatformInterface extends PlatformInterface {
   Future<void> capture({
     required String eventName,
     Map<String, Object>? properties,
+    Map<String, Object>? groups,
   }) {
     throw UnimplementedError('capture() has not been implemented.');
   }